Rename TryEx to TryTo, credits to @norzechowicz
