+++ src/Control/TryTo/Success.php
@@ -2,15 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Munus\Control\TryEx;
+namespace Munus\Control\TryTo;
 
-use Munus\Control\TryEx;
+use Munus\Control\TryTo;
 
 /**
  * @template T
  * @template-extends TryEx<T>
  */
-final class Success extends TryEx
+final class Success extends TryTo
 {
     /**
      * @var T
