+++ tests/Control/TryToTest.php
@@ -4,17 +4,17 @@ declare(strict_types=1);
 
 namespace Munus\Tests\Control;
 
-use Munus\Control\TryEx;
+use Munus\Control\TryTo;
 use Munus\Tests\Stub\Expect;
 use Munus\Tests\Stub\Result;
 use Munus\Tests\Stub\Success;
 use PHPUnit\Framework\TestCase;
 
-final class TryExTest extends TestCase
+final class TryToTest extends TestCase
 {
     public function testTrySuccess(): void
     {
-        $try = TryEx::of(function () {return 'success'; });
+        $try = TryTo::run(function () {return 'success'; });
 
         self::assertTrue($try->isSuccess());
         self::assertEquals('success', $try->get());
@@ -22,7 +22,7 @@ final class TryExTest extends TestCase
 
     public function testTryFailure(): void
     {
-        $try = TryEx::of(function () {throw new \DomainException('use ddd'); });
+        $try = TryTo::run(function () {throw new \DomainException('use ddd'); });
 
         self::assertTrue($try->isFailure());
         self::assertEquals(new \DomainException('use ddd'), $try->getCause());
@@ -30,8 +30,8 @@ final class TryExTest extends TestCase
 
     public function testGetOrSomething(): void
     {
-        /** @var TryEx<Result> $try */
-        $try = TryEx::of(function () {throw new \DomainException('use ddd'); });
+        /** @var TryTo<Result> $try */
+        $try = TryTo::run(function () {throw new \DomainException('use ddd'); });
         $result = $try->getOrElse(new Result());
 
         Expect::result($result);
@@ -42,26 +42,26 @@ final class TryExTest extends TestCase
     public function testEquals(): void
     {
         $exception = new \DomainException('use ddd');
-        self::assertTrue(TryEx::of(function (): Success {return new Success(); })->equals(new Success()));
-        self::assertTrue(TryEx::of(function () use ($exception) {throw $exception; })->equals($exception));
+        self::assertTrue(TryTo::run(function (): Success {return new Success(); })->equals(new Success()));
+        self::assertTrue(TryTo::run(function () use ($exception) {throw $exception; })->equals($exception));
     }
 
     public function testMapFailure(): void
     {
-        $try = TryEx::of(function () {throw new \DomainException('use ddd'); });
+        $try = TryTo::run(function () {throw new \DomainException('use ddd'); });
         self::assertEquals(new \DomainException('use ddd'), $try->map('strtolower')->getCause());
     }
 
     public function testMapWithMapperFailure(): void
     {
-        $try = TryEx::of('time');
+        $try = TryTo::run('time');
 
         self::assertInstanceOf(\TypeError::class, $try->map('strtoupper')->getCause());
     }
 
     public function testMapWitMapperSuccess(): void
     {
-        $try = TryEx::of(function () {return 'success'; });
+        $try = TryTo::run(function () {return 'success'; });
 
         self::assertEquals('SUCCESS', $try->map('strtoupper')->get());
     }
